backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
